+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Unreleased
 
-* Drop support for python 3.8 (\#438).
 * Update `_TaskBase` schema from `fractal-common` (\#431).
+* Drop support for python 3.8 (\#438).
+* Update `fractal task edit` command (\#439).
 * Disable `argparse` abbreviation for CLI commands (\#441).
 
 # 1.0.5